update ProteinView with new tab views
